v50-51
git-svn-id: svn://192.168.1.10/public/CoffeeMud@5397 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Commands/BaseClanner.java
@@ -137,7 +137,7 @@ public class BaseClanner extends StdCommand
 		String[] roles=Clan.ROL_DESCS[government];
 		for(int i=0;i<roles.length;i++)
 			if(roles[i].startsWith(position.toUpperCase()))
-				return CMath.pow(2,i-1);
+				return (int)CMath.pow(2,i-1);
 		return -1;
 	}
